backend/requirements.txt
fastapi
uvicorn
pydantic
pandas
pyjwt
passlib[bcrypt]
motor
python-dotenv
